+++ CHANGELOG.md
@@ -4,3 +4,4 @@
   fields instead
 * `natura_iva` is now from 2 to 4 characters long (#18)
 * Added a38.consts module with constants for common enumerations (#18)
+* Added `DettaglioLinee.autofill_prezzo_totale`